Root: Remove nc & use docker healthchecks
We don't need to test connection with netcat, instead we can use docker healthchecks to check the health status of a container.
You can run docker ps to check the status: should say (healthy), if (unhealthy), it never builds flow in the first place.
You could build an autoheal system, but I think this is enough.