main.js/changeContainerContent: Update content with DocumentFragment
Using the .innerHTML to update the page contents considers the
contents as plain text, so the browser can not comprehend the HTML
elements we are inserting. The big problem with that was that the
page-specific <script> tags would not get executed. This new method
loads the content to DocumentFragment objects before replacing it,
which solves that problem, and shold be more robust solution
altogether.