pong.js: Add function to set up click events for page links
The function adds click event listener links that navigate the
application to another sub-page. Those links will be identified from
'page_link' class.

Clicking will get the content from the href-attribute of the link,
and the attribute should be prepended with '#' to avoid accidentally
linking to another (broken) page.

For example, clicking the following gets the page from /content/page:
<a class="page_link" href="#page">link to page</a>