flow: Fix redirect issue for authenticated users
Previously the logged users were incorrectly directed to the landing page
when refreshing. The authentication was being checked inside the index.html
file. This was fixed by checking for user authentication in the index endpoint
which renders the template. If it is authenticated, it directs them to the home page.