flow: Use location hashes for page navigation
Load page content based on location hash value, and update it on page
load and hash changes. This enables back/forward button functionality
and linking directly to a specific page.

This is also more elegant implementation than before, since we don't
need to mark page-changing links with the weird 'page_link' class
and remember to use the initPageLinks() function on every page.
Simply linking to an anchor, or updating the location hash value
works from now on.