landing.html: Display continue button if user is logged in
With old page navigation scheme, user could only access the landing
page if they were not logged in. Since this is now changed, we need
to display something that makes sense for logged in users, so
replace the login and register buttons with something that allows
them to continue to home page.