register.html,login.html: Enclose scripts inside block statements
This prevents the locally used varibles from being declared in
global namespace. This was causing problems before whenever the
pages were accessed second time, which caused the form variables
to get re-declared, which is not alloved for const variables.
Using global variables is not generally recommended either way.

Group the element look-up variable declarations together while at
it for cleaner look, since it is now possible.