game.js: Thin paddle rectangles by ball radius
Whether the ball hits the paddle is measured from the center point
of the ball on the web game UI. This caused a issue where the ball
appeared to go trough the paddle edge, as the ball center was
outside of the paddle edge, but the apparent ball radius still was
within the paddle area. Fix this by making the paddle boxes one ball
radius thinner on both ends.