onlineMatchConsumer: Stop sending start_match twice
This causes issues on web game as it causes the game to be
initialized twice. IIRC this was dirty fix for some issue, but it
is not present in any other consumers. If the issue that this worked
around still exists, there must be a more proper solution for it.