templates: Add autofocus in login & register
I tested with the html attribute 'autofocus' but it didn't work when the user logged out by themselves and was prompted the login screen: it wouldn't autofocus. Also in the lobby I tried the hmtl tags but when the user just reigstered their account, the attribute in html wouldn't focus on the chat. That's why I had to use javascript for this, even though I would have liked to use the html attributes.