onlineTournamentConsumer.py & ws.js: Add player disconnect message
I need to save the username to self.userame in the connect, because this is still in the lobby, and so we can't use the self.player1Username because those are initialized only when the game has begun. I feel like this whole thing could be restructured heavily but we are short on time. Also I don't think we need to decrement the self.connectedPlayers variable on disconnection, because our backend doesn't actually handle that, like we have discussed, we just mark them as 'disconnected' and the players themselves can decide what the fuck they'll do about it. I tried to do it the right way for some hours, but then remembered that restructuring much of the back, would prolly break the cli_client as well, so we are stuck with the tape fix.