lobby.html: Set default speeder value
This was once removed, but I am going to add it back, because for local games we never send the speeds to server, hence we are not getting updates from there either. We keep the speeder value localized in local games, that's why it makes sense to set the default here. On online games it gets overwritten.